Lots of Additions
Boss bar effect now supports "variable-max" for a variable as the maximum. This allows boss bars to be much more dynamic.

Added a passive trigger for "startswim" and "stopswim" which triggers when you start and stop swimming. Allows you to do things like speed up swimming or add effects to it.

Valid target now supports "mounts", "passengers", "castermount", and "casterpassenger" to allow you to target specific entities in stacks.

Added a meta variable that allows you to grab the current timestamp in days, hours, minutes, and seconds. This is a huge step forward for scheduling spells / time measurement.